# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "The Reward Loyalty API provides programmatic access to the Reward Loyalty platform, enabling integration with external systems, mobile apps, and custom workflows. This RESTful API supports authentication, loyalty card management, member operations, partner operations, and transaction processing.\n\n## Authentication\n\nAll authenticated endpoints use Bearer token authentication via Laravel Sanctum. Include your token in the Authorization header.\n\nTokens are obtained by calling the login endpoints for each user type (Admin, Partner, or Member).\n\n## Rate Limiting\n\nAPI requests are rate-limited to ensure fair usage. Standard limits apply per authenticated user.\n\n## Response Format\n\nAll responses are returned in JSON format. Successful responses include the requested data, while errors include a message field describing the issue.",
        "contact": {
            "name": "Reward Loyalty Website",
            "url": "https://rewardloyalty.co"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Current Server"
        }
    ],
    "paths": {
...
        "/{locale}/v1/admin/partners": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List all accessible partners",
                "description": "Retrieve a list of all partners the authenticated administrator has access to. Administrators (role=1) can view all partners; Managers see only partners within their assigned networks.",
                "operationId": "getAdminPartners",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of partners",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdminPartner"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "admin_auth_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create a new partner",
                "description": "Create a new partner account with optional permissions and limits. A default club is automatically created for the partner.",
                "operationId": "createAdminPartner",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Partner data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "network_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Partner name",
                                        "type": "string",
                                        "example": "Coffee Corner"
                                    },
                                    "email": {
                                        "description": "Partner email",
                                        "type": "string",
                                        "format": "email",
                                        "example": "partner@example.com"
                                    },
                                    "password": {
                                        "description": "Password (6-48 chars)",
                                        "type": "string",
                                        "format": "password",
                                        "example": "securepass123"
                                    },
                                    "network_id": {
                                        "description": "Network ID to assign partner to",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "locale": {
                                        "description": "Locale",
                                        "type": "string",
                                        "example": "en_US"
                                    },
                                    "currency": {
                                        "description": "Currency code",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "time_zone": {
                                        "description": "Time zone",
                                        "type": "string",
                                        "example": "America/New_York"
                                    },
                                    "is_active": {
                                        "description": "Active status",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "permissions": {
                                        "$ref": "#/components/schemas/PartnerPermissions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Partner created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminPartnerFull"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "admin_auth_token": []
                    }
                ]
            }
        },
...
        "/{locale}/v1/partner": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "Retrieve authenticated partner's data",
                "description": "Retrieve the data of the authenticated partner.",
                "operationId": "getPartner",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Partner data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Partner"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
...
        "/{locale}/v1/partner/": {
            "put": {
                "tags": [
                    "Partner"
                ],
                "summary": "Update the authenticated partner's details",
                "description": "Update the profile details of the currently authenticated partner. Only include fields you wish to modify.",
                "operationId": "updatePartner",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "The locale (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Provide the partner data you wish to update. Omit any key/value pairs that should remain unchanged.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Partner's name",
                                        "type": "string",
                                        "example": "Partner Name"
                                    },
                                    "email": {
                                        "description": "Partner's email",
                                        "type": "string",
                                        "example": "partner@example.com"
                                    },
                                    "locale": {
                                        "description": "Locale setting",
                                        "type": "string",
                                        "example": "en-us"
                                    },
                                    "currency": {
                                        "description": "Currency setting",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "time_zone": {
                                        "description": "Time zone setting",
                                        "type": "string",
                                        "example": "America/Los_Angeles"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Partner updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Partner"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "description": "Response when authentication is required but not provided or invalid",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "Authentication error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "description": "Response when request validation fails",
                "properties": {
                    "message": {
                        "description": "General validation error message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Field-specific validation errors",
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "AdminPartner": {
                "description": "Partner data as viewed by an administrator",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier",
                        "type": "string",
                        "format": "uuid",
                        "example": "019b8d48-97df-7177-8aa0-27f2107a8eb1"
                    },
                    "network_id": {
                        "description": "Associated network identifier",
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "name": {
                        "description": "Business name",
                        "type": "string",
                        "maxLength": 64,
                        "example": "Coffee Corner"
                    },
                    "email": {
                        "description": "Contact email",
                        "type": "string",
                        "format": "email",
                        "example": "partner@coffecorner.com"
                    },
                    "email_verified_at": {
                        "description": "Email verification timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "locale": {
                        "description": "Preferred locale",
                        "type": "string",
                        "example": "en_US",
                        "nullable": true
                    },
                    "currency": {
                        "description": "Default currency",
                        "type": "string",
                        "example": "USD",
                        "nullable": true
                    },
                    "time_zone": {
                        "description": "Time zone",
                        "type": "string",
                        "example": "America/New_York",
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Whether the partner account is active",
                        "type": "boolean",
                        "example": true
                    },
                    "number_of_times_logged_in": {
                        "description": "Total login count",
                        "type": "integer",
                        "example": 15
                    },
                    "last_login_at": {
                        "description": "Last login timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Account creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last modification timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "avatar": {
                        "description": "URL to avatar image",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "AdminPartnerFull": {
                "description": "Partner data with permissions as viewed by an administrator",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/AdminPartner"
                    },
                    {
                        "properties": {
                            "permissions": {
                                "$ref": "#/components/schemas/PartnerPermissions"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "admin_auth_token": {
                "type": "http",
                "description": "Administrator authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "partner_auth_token": {
                "type": "http",
                "description": "Partner authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Member authentication token obtained from the login endpoint. Include in the Authorization header as `Bearer <token>`.",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}